requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Automat==25.4.16
certifi==2026.2.25
cffi==2.0.0
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
cssselect==1.4.0
defusedxml==0.7.1
fastapi==0.135.1
filelock==3.25.1
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
jmespath==1.1.0
lxml==6.0.2
lxml_html_clean==0.4.4
packaging==26.0
parsel==1.11.0
Protego==0.6.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyDispatcher==2.0.7
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
queuelib==1.9.0
readability-lxml==0.8.4.1
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
Scrapy==2.14.1
service-identity==24.2.0
six==1.17.0
starlette==0.52.1
tldextract==5.3.1
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
w3lib==2.4.0
zope.interface==8.2
